pyproject.toml
[project]
name = "one-billion-row-challenge-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "dask-expr>=1.1.20",
    "dask[complete]>=2024.12.0",
    "duckdb>=1.1.3",
    "modin[dask]>=0.32.0",
    "pandas>=2.2.3",
    "polars>=1.17.1",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
